move Post to dto package + make fun changeCounters private
