+++ app/src/main/java/ru/netology/nmedia/dto/Post.kt
@@ -1,0 +1,11 @@
+package ru.netology.nmedia.dto
+
+data class Post(
+    val id: Long,
+    val author: String,
+    val content: String,
+    val published: String,
+    val likedByMe: Boolean,
+    val counterMap: MutableMap<String, Long>,
+    val title: String
+)
