+++ app/src/main/java/ru/netology/nmedia/repository/PostRepository.kt
@@ -1,7 +1,7 @@
 package ru.netology.nmedia.repository
 
 import androidx.lifecycle.LiveData
-import ru.netology.nmedia.ui.Post
+import ru.netology.nmedia.dto.Post
 
 interface PostRepository {
     fun get(): LiveData<Post> // получить объект Post
