+++ app/src/main/java/ru/netology/nmedia/ui/MainActivity.kt
@@ -45,13 +45,3 @@ class MainActivity : AppCompatActivity() {
 
 }
 
-data class Post(
-    val id: Long,
-    val author: String,
-    val content: String,
-    val published: String,
-    val likedByMe: Boolean,
-    val counterMap: MutableMap<String, Long>,
-    val title: String
-)
-
